table_data_extraction/requirements.txt
absl-py==1.4.0
addict==2.4.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrdict==2.0.1
attributee==0.1.7
attrs==22.2.0
Babel==2.11.0
bce-python-sdk==0.8.79
beautifulsoup4==4.11.2
black==23.1.0
Brotli==1.0.9
cachelib==0.10.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cityscapesScripts==2.2.1
click==8.1.3
cloudpickle==2.2.1
codecov==2.1.12
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.0.7
coverage==7.1.0
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.33
decorator==5.1.1
dill==0.3.6
dotty-dict==1.3.1
einops==0.6.0
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastapi==0.92.0
ffmpy==0.3.0
filelock==3.9.0
fire==0.5.0
flake8==6.0.0
Flask==2.2.2
flask-babel==3.0.1
Flask-Session==0.4.0
flatbuffers==23.1.21
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.4.0
gevent==22.10.2
geventhttpclient==2.0.2
google-api-core==2.11.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-pasta==0.2.0
googleapis-common-protos==1.58.0
gradio==3.18.0
greenlet==2.0.2
grpcio==1.51.1
h11==0.14.0
h5py==3.8.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.12.0
humanfriendly==10.0
hydra-core==1.3.1
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
importlib-resources==5.10.2
iniconfig==2.0.0
interrogate==1.5.0
iopath==0.1.9
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
kiwisolver==1.4.4
kwarray==0.6.9
lap==0.4.0
libclang==15.0.6.1
linkify-it-py==1.0.3
lmdb==1.4.0
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.5.2
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mmcls==0.25.0
mmdet==2.28.1
motmetrics==1.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.0
numpy==1.23.0
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.7.0
onnxruntime==1.14.0
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
opencv-python-headless==4.6.0.66
openpyxl==3.1.1
opt-einsum==3.3.0
orjson==3.8.6
packaging==23.0
paddle-bfloat==0.1.7
paddleocr==2.6.0.1
paddlepaddle==2.3.2
pandas==1.4.3
pathspec==0.11.0
pdf2docx
pdf2image==1.16.0
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
platformdirs==3.0.0
pluggy==1.0.0
portalocker==2.7.0
premailer==3.10.0
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
pycodestyle==2.10.0
pycryptodome==3.17
pydantic==1.10.4
pydub==0.25.1
pyflakes==3.0.1
pymongo==4.1.1
PyMuPDF==1.20.2
pyocr==0.8.3
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.19.3
pytesseract==0.3.9
pytest==7.2.1
python-dateutil==2.8.2
python-docx==0.8.11
python-multipart==0.0.5
python-rapidjson==1.9
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.13.7
rarfile==4.0
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
seaborn==0.12.2
shapely==2.0.1
six==1.16.0
scikit-learn
sniffio==1.3.0
soupsieve==2.4
starlette==0.25.0
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminaltables==3.1.10
tifffile==2023.2.3
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torchvision==0.14.1
tqdm==4.64.1
transformers==4.20.1
tritonclient==2.30.0
typing==3.7.4.3
typing-extensions==4.5.0
ubelt==1.2.3
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
visualdl==2.5.0
websockets==10.4
Werkzeug==2.2.3
wrapt==1.14.1
x2paddle==1.4.0
xdoctest==1.1.1
xmltodict==0.13.0
yacs==0.1.8
yapf==0.32.0
yarl==1.8.2
zipp==3.13.0
zope.event==4.6
zope.interface==5.5.2